scraper-cz/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scraper-cz"
version = "0.1.0"
description = "Scraper for Czech National Archives VadeMeCum portal"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27",
    "Pillow>=10.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "responses>=0.25",
]

[project.scripts]
scraper-cz = "scraper_cz.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper_cz"]
